Split app.js phase 4a: extract schedule-editor and schedule-overlay modules
Schedule segments/repeat/overlay-selection state moved to state.js. In-app
tests pass (132/132).
