+++ src/state.js
@@ -45,4 +45,11 @@ export const state = {
     activatedBlockIds: new Set(), // Track blocks that have already triggered host updates
     pauseScheduleData: null, // Track schedule-specific pause data { blocklistId, segmentEndTime }
     isScheduleMode: false, // false = instant mode, true = schedule mode
+    scheduleSegments: null, // Array of time segments with per-segment days (set at startup)
+    expandedScheduleSegmentIndex: 0, // Which segment shows the full editor when multiple exist (-1 = all collapsed)
+    scheduleRepeatDate: null, // Date object when repeatType is 'date'
+    mobileCompactScheduleDayLabelsActive: null,
+    appBlockingActiveStartOverlay: null, // Custom start-overlay config for the active schedule warning session
+    pendingScheduleStartOverlayId: null,
+    scheduleOverlayCustomiseSelection: null,
 };
